added EnMAPBoxAbstractContextMenuProvider implemented first methods in EnMAPBoxContextMenuProvider populateMapCanvasContextMenu populateDataViewContextMenu refactoring
